fix(15.11): Unblock app shell while platform features load.
Avoid a blank screen during feature fetch and add MSW handlers so auth routing tests reach the dashboard when onboarding is disabled.
