+++ clients/web/src/components/onboarding/onboarding-redirect.tsx
@@ -1,47 +1,6 @@
-import { useEffect, useState, type ReactNode } from 'react'
-import { Navigate, useLocation } from 'react-router-dom'
-import { getAccountType } from '../../lib/auth'
-import { fetchOnboardingStatus } from '../../lib/onboarding-api'
-import { usePlatformFeatures } from '../../context/platform-features-context'
-
-/** Redirects new self-learners to onboarding when the feature flag is on and flow is incomplete. */
-export function useOnboardingRedirect(): { checking: boolean; shouldRedirect: boolean } {
-  const { ffOnboardingFlow, loading: featuresLoading } = usePlatformFeatures()
-  const location = useLocation()
-  const [checking, setChecking] = useState(true)
-  const [shouldRedirect, setShouldRedirect] = useState(false)
-
-  useEffect(() => {
-    if (featuresLoading) return
-    if (!ffOnboardingFlow || getAccountType() === 'parent') {
-      setShouldRedirect(false)
-      setChecking(false)
-      return
-    }
-    if (location.pathname.startsWith('/onboarding')) {
-      setShouldRedirect(false)
-      setChecking(false)
-      return
-    }
-    let cancelled = false
-    void fetchOnboardingStatus()
-      .then((status) => {
-        if (cancelled) return
-        setShouldRedirect(Boolean(status && !status.completed))
-      })
-      .catch(() => {
-        if (!cancelled) setShouldRedirect(false)
-      })
-      .finally(() => {
-        if (!cancelled) setChecking(false)
-      })
-    return () => {
-      cancelled = true
-    }
-  }, [featuresLoading, ffOnboardingFlow, location.pathname])
-
-  return { checking, shouldRedirect }
-}
+import type { ReactNode } from 'react'
+import { Navigate } from 'react-router-dom'
+import { useOnboardingRedirect } from './use-onboarding-redirect'
 
 export function OnboardingRedirect({ children }: { children: ReactNode }) {
   const { checking, shouldRedirect } = useOnboardingRedirect()
